kumo_server_common::authn_authz

Struct AuthInfo

pub struct AuthInfo {
    pub peer_address: Option<IpAddr>,
    pub identities: Vec<Identity>,
    pub groups: Vec<String>,
}

Fields
peer_address: Option<IpAddr>

The peer machine, if known

identities: Vec<Identity>

Authenticated identities

groups: Vec<String>

Any groups to which we might belong

Implementations

impl AuthInfo

pub fn new_local_system() -> Self

Authentication that represents the local system taking action. It should be considered to be the most privileged identity

pub fn summarize_for_http_auth(&self) -> String

pub fn public_descriptor(&self) -> String

Summarize the info in a form that should be reasonable to report back to a connected peer. This form doesnโt include any groups that may have been added by the server and/or via policy

pub fn merge_from(&mut self, other: Self)

Merges groups and identities from other, ignoring its peer_address

pub fn add_identity(&mut self, identity: Identity)

Add an identity to the list

pub fn add_group(&mut self, group_name: impl Into<String>)

pub fn set_peer_address(&mut self, peer_address: Option<IpAddr>)

Set the peer address, ignoring and replacing any prior value

pub fn matches_identity(&self, identity: &ACLIdentity) -> bool

Test to see if this AuthInfo matches a specific identity from an ACL entry

pub fn matches_criteria(&self, criteria: &ACLRuleTerm) -> bool