kumo_server_common::http_server

Module auth

Structs

TrustedIpRequired

Use this type as an extractor parameter when the handler must only be accessible to trusted IP addresses

Enums

AuthKind

Represents some authenticated identity. Use this as an extractor parameter when you need to reference that identity in the handler.

Functions

auth_middleware