kumo_server_common::log_backlog

Function register_location

pub fn register_location(location: WriterLocation)

Add location to the monitored set and start the monitor if it is not already running. Registering the same location more than once, including when a configuration reload re-registers the same directory, has the same effect as registering it once. A location is monitored for the lifetime of the process. A location cannot be deregistered.